python/07-监控与可观测性/examples/requirements.txt
# Python 监控示例应用依赖
# Python 3.12+

# Web框架
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.0

# 监控
prometheus-client==0.21.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp==1.27.0

# 日志
structlog==24.4.0

# 工具
python-multipart==0.0.9

